api/requirements.txt
# FastAPI and dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Databricks SDK
databricks-sdk==0.18.0

# HTTP client
requests==2.31.0

# Python utilities
python-multipart==0.0.6
python-dotenv==1.0.0

# Development dependencies (optional)
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
